Added copy-files mechanism to Context, and c_add_modules now copies files
The ThingMLCompiler class now keeps track of input directory so that it
can be used (also by the context) to find files relative to the .thingml
file that was given to the compiler.

The Context class has a map of files that is supposed to be copied to
the output directory, and it is implemented for the Posix and Posix-MT
compilers, and used with the "add_c_modules" annotation (if it can find
a .c and .h file with the same name as the module in the input
directory).