Added specification of topic for PosixMQTTPlugin
The topic to publish messages to can be specified by an annotation on the
protocol instance or on external connectors.

Currently, the subscribed topics can only be configured globally for the
protocol instance. This means that any topics defined on an external
connector, will also affect other external connectors over the same
protocol.

To fix this, the C compilers need to be modified so that external messages
can be enqeued with a destination port.