Removed some references to model.eResource()
When models are flattened, they are not put into a new resource (I tried
to make some "in-memory" resource, but EMF wouldn't allow me - which might
make sense). So if you try to re-flatten the model, it fails. And when the
checker tries to find the source file, it fails.

It works now, but I'm not completely sure when we actually need to use the
ResolveAll() which is now disabled if the model has no resource.