First pass at brand new ThingML testing framework
The new framework runs the old ThingML tests using JUnit and has been
tested using both Eclipse and Maven (on Windows).

The ThingML "API" has changed, so all the tests have been updated, which
might mean that some of the tests have actually been broken. I guess it is
time to go through them anyways...

Currently, the NodeJS and Maven compilers are supported.

It probably needs a few more passes, and some refactoring to make it nice,
but all the functionality (to run the tests on your local machine) should
be in place :)