Refactoring of the new testing framework
Some parts have been rewritten to allow for multithreaded execution, and
for preparation of remote test execution.

The plan for remote test execution is that the tests are provided and
prepared (the parts that take place inside the JVM - and are quick) on the
local machine, resulting in instances of the ThingMLTest class. These
instances (which are serializable) will be distributed to workers somehow,
which in turn re-loads these instances, generates the ThingMLTestCase
instances, and runs the test. The results of the test execution should
then be reported back (probably in form of serialized AssertionErrors),
and reported as normal on the local machine.