Added new back-tick terminal in the grammar
This type of string `...` can be used in both annotations and external
statements and expressions, should fix some issues with escaping in
annotations, and make it easier to use single quotes in externals.

The back-ticks are the same syntax as Markdown uses for code snippets,
which should be familiar to lot's of people.

It is also a step towards using '.' for characters in ThingML.

Fixes #186
Related to #156