New ThingML build support in Eclipse
The new system uses LaunchConfigurations and the provided tools to store
the configuration, and the history of builds. Some tweaks/hacks are
implemented since we don't really have Run/Debug configurations, but
rather Generate (and in the future, Run) configurations. But they at
least work well in Oxygen 3A.

The ThingML workspace has to be selected to show the new features.