Fix global warming due to ThingML
A state machine with a final state has an infinite loop for every new incoming message.
It's because the state machine is stopped, and therefore not ready, so the message is sent again at the next tick.
If the state machine has a final state and uses timers, it will receive quite a lot of messages that will loop forever, using more and more cpu.